TSO_Search_Env , WS_SO_Search   - Mudado tipo da variavel product_code
Act021, act022
  - Mudado o nome das constantes utilizadas no bundle.

 Act023
  - Criado metodos que buscam os dados do produto
  - Criado metodo que valida se serial esta vazio
  - Criado metodos que definem o fluxo da informação , baseado no requesting_process.

ConstanteBaseApp
  - Add constantes

Res
  - Modificado Layout das act 023 e 024

+++ app/src/main/java/com/namoadigital/prj001/model/TSO_Search_Env.java
@@ -9,7 +9,7 @@ public class TSO_Search_Env {
     private String app_code;
     private String app_version;
     private String session_app;
-    private int product_code;
+    private Long product_code;
     private String serial_id;
     private String so_mult;
 
@@ -37,11 +37,11 @@ public class TSO_Search_Env {
         this.session_app = session_app;
     }
 
-    public int getProduct_code() {
+    public Long getProduct_code() {
         return product_code;
     }
 
-    public void setProduct_code(int product_code) {
+    public void setProduct_code(Long product_code) {
         this.product_code = product_code;
     }
